Make the “today” and “history” cards public
