+++ SwiftPackage/Sources/BridgeClientUI/SingleStudy/Views/HistoryView.swift
@@ -33,7 +33,7 @@ public struct HistoryView: View {
                     CustomScrollView {
                         LazyVStack(spacing: 16) {
                             ForEach(viewModel.records) { record in
-                                card(record)
+                                HistoryCardView(record)
                             }
                         }
                         .padding(.bottom, 20)
@@ -75,52 +75,7 @@ public struct HistoryView: View {
         CompositeTintedImage("History.tinted", background: "History", bundle: .module)
     }
     
-    @ViewBuilder
-    private func card(_ record: AssessmentRecord) -> some View {
-        HStack(alignment: .top, spacing: 0) {
-            rightSide(record)
-            leftSide(record)
-        }
-        .padding(.horizontal, 31)
-    }
-    
-    @ViewBuilder
-    private func rightSide(_ record: AssessmentRecord) -> some View {
-        assessmentInfoMap.color(for: record.info)
-            .overlay(assessmentInfoMap.icon(for: record.info))
-            .frame(width: 114)
-            .frame(minHeight: 112)
-    }
-    
-    let detailHeaderFont = DesignSystem.fontRules.headerFont(at: 7)
-    let titleHeaderFont = DesignSystem.fontRules.headerFont(at: 5)
-    
-    @ViewBuilder
-    private func leftSide(_ record: AssessmentRecord) -> some View {
-        VStack(alignment: .leading, spacing: 2) {
-            CompletedDotView(color: assessmentInfoMap.color(for: record.info))
-                .padding(.bottom, 2)
-            Text(DateFormatter.string(from: record.finishedOn, timeZone: record.timeZone, dateStyle: .medium, timeStyle: .none))
-                .foregroundColor(.textForeground)
-                .font(detailHeaderFont)
-            assessmentInfoMap.title(for: record.info)
-                .font(titleHeaderFont)
-                .foregroundColor(.textForeground)
-                .frame(maxWidth: .infinity, alignment: .leading)
-            Text(durationFormatter.string(from: Double(record.minutes * 60)) ?? "")
-                .foregroundColor(.textForeground)
-                .font(detailHeaderFont)
-        }
-        .padding(.leading, 16)
-        .padding(.top, 8)
-    }
-    
-    private let durationFormatter: DateComponentsFormatter = {
-        let formatter = DateComponentsFormatter()
-        formatter.unitsStyle = .short
-        formatter.allowedUnits = .minute
-        return formatter
-    }()
+
 }
 
 extension DateFormatter {
